requirements.txt
fastapi
uvicorn[standard]
playwright
beautifulsoup4
google-generativeai
chromadb
sentence-transformers
torch
python-dotenv
pyttsx3
numpy
vosk
protobuf==3.20.*
streamlit
pillow
gTTS
openai-whisper
pydub
audio_recorder_streamlit